重构: 将用于订阅流的文章结构体提取独立模块
