+++ libbmy/CMakeLists.txt
@@ -5,6 +5,7 @@ add_library(bmy SHARED
 	cookie.c
 	wechat.c
 	convcode.c
+	article.c
 )
 
 target_link_libraries(bmy ytht mysqlclient ssl crypto curl json-c)
